# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZeeoWork API Name",
        "description": "ZeeoWork API Description",
        "contact": {
            "email": "tech@zeework.ir"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://devpandora.zeework.ir/"
        }
    ],
    "paths": {
...
        "/api/hr/employees": {
            "get": {
                "tags": [
                    "OrganizationAdmin/Employees"
                ],
                "summary": "Get a paginated list of employees for the organization",
                "operationId": "3cf0f9346b1838be942219ce5fb167ef",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for employee name, national code, or email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "Filter employees by a specific department ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "first_name",
                                "last_name",
                                "national_code",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AdminEmployeeResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "OrganizationAdmin/Employees"
                ],
                "summary": "OA docs for this route",
                "operationId": "03f1a536ad07a990f7e5691225f4fd30",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrganizationAdminEmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationAdminEmployeeResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hr/employees/{id}": {
            "get": {
                "tags": [
                    "OrganizationAdmin/Employees"
                ],
                "summary": "OA docs for this route",
                "operationId": "07a409d5d2cdb3ec2c9483d1db886bb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrganizationAdmin ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationAdminEmployeeResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "OrganizationAdmin/Employees"
                ],
                "summary": "OA docs for this route",
                "operationId": "e6a416864874f198d05b17110d6fffca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrganizationAdmin ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/schemas/OrganizationAdminEmployeeRequest"
                },
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "OrganizationAdmin/Employees"
                ],
                "summary": "OA docs for this route",
                "operationId": "e10fbf3062a64bfa4d9f0b646de038aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrganizationAdmin ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hr/employees/import": {
            "post": {
                "tags": [
                    "OrganizationAdmin/Employees"
                ],
                "summary": "Import employees from an Excel file for the authenticated admin's organization",
                "operationId": "776c849a90feebceca47d4cd9f1e6b45",
                "requestBody": {
                    "description": "Excel file containing employee data. The file must have a header row with columns: `national_code`, `name`, `email`, `phone`, `job_position`, `departments`. The `departments` column should contain comma-separated department names.",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "The .xlsx or .csv file to import.",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import was successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error, for example, if the file is missing or has validation errors in rows.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": true
                    },
                    "messages": {
                        "description": "آرایه‌ای از پیام‌ها برای کاربر (می‌تواند شامل پیام‌های موفقیت، خطا یا اطلاعاتی باشد)",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "عملیات با موفقیت انجام شد."
                        },
                        "example": [
                            "عملیات موفقیت‌آمیز بود",
                            "پیام دوم در صورت وجود"
                        ],
                        "nullable": true
                    },
                    "data": {
                        "type": "object",
                        "example": {
                            "id": 1,
                            "name": "Ali"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "OrganizationAdminEmployeeRequest": {
                "title": "Organization Admin Employee Request",
                "required": [
                    "national_code",
                    "name",
                    "email",
                    "phone",
                    "department_ids",
                    "job_position"
                ],
                "properties": {
                    "national_code": {
                        "description": "کد ملی کارمند (۸ تا ۱۰ رقم، یکتا)",
                        "type": "string",
                        "example": "0071234567"
                    },
                    "name": {
                        "description": "نام کامل کارمند",
                        "type": "string",
                        "maxLength": 255,
                        "example": "علی رضایی"
                    },
                    "email": {
                        "description": "ایمیل کارمند (یکتا)",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "ali.rezaei@example.com"
                    },
                    "phone": {
                        "description": "شماره موبایل کارمند (یکتا)",
                        "type": "string",
                        "example": "09123456789"
                    },
                    "photo": {
                        "description": "عکس پرسنلی (jpeg, png, jpg, gif - حداکثر ۲ مگابایت، اختیاری)",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    },
                    "department_ids": {
                        "description": "آرایه آیدی دپارتمان‌ها (حداقل یک مورد، فقط دپارتمان‌های این سازمان)",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 2
                        },
                        "minItems": 1
                    },
                    "job_position": {
                        "description": "سمت شغلی کارمند",
                        "type": "string",
                        "maxLength": 100,
                        "example": "برنامه‌نویس ارشد"
                    }
                },
                "type": "object"
            },
...
            "AdminEmployeeResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "national_code": {
                        "type": "string",
                        "example": "0071234567"
                    },
                    "name": {
                        "type": "string",
                        "example": "علی رضایی"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ali.rezaei@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "09123456789"
                    },
                    "photo": {
                        "type": "string",
                        "example": "https://example.com/uploads/employees/123.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
            "OrganizationAdminDepartmentResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "واحد مالی"
                    },
                    "description": {
                        "type": "string",
                        "example": "این دپارتمان مسئول امور مالی شرکت است.",
                        "nullable": true
                    },
                    "employee_count": {
                        "description": "تعداد کارمندان دپارتمان (در صورت بارگذاری relation)",
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01 08:30:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-10 11:21:00"
                    }
                },
                "type": "object"
            },
            "OrganizationAdminEmployeeResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 23
                    },
                    "national_code": {
                        "type": "string",
                        "example": "0071234567"
                    },
                    "name": {
                        "type": "string",
                        "example": "علی رضایی"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ali.rezaei@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "09123456789"
                    },
                    "photo": {
                        "type": "string",
                        "format": "url",
                        "example": "https://cdn.example.com/employees/23.jpg",
                        "nullable": true
                    },
                    "job_position": {
                        "type": "string",
                        "example": "برنامه‌نویس ارشد",
                        "nullable": true
                    },
                    "departments": {
                        "description": "لیست دپارتمان‌های کارمند در این سازمان",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrganizationAdminDepartmentResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01 14:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-02 09:30:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}